[ANDROSDK-1964] Move CategoryFields to network package
