+++ core/src/main/java/org/hisp/dhis/android/core/category/CategoryCollectionRepository.kt
@@ -34,7 +34,7 @@ import org.hisp.dhis.android.core.arch.repositories.filters.internal.FilterConne
 import org.hisp.dhis.android.core.arch.repositories.filters.internal.StringFilterConnector
 import org.hisp.dhis.android.core.arch.repositories.scope.RepositoryScope
 import org.hisp.dhis.android.core.category.internal.CategoryCategoryOptionChildrenAppender
-import org.hisp.dhis.android.core.category.internal.CategoryFields
+import org.hisp.dhis.android.network.category.CategoryFields
 import org.hisp.dhis.android.core.category.internal.CategoryStore
 import org.koin.core.annotation.Singleton
