+++ core/src/main/java/org/hisp/dhis/android/network/category/CategoryNetworkHandlerImpl.kt
@@ -27,12 +27,10 @@
  */
 package org.hisp.dhis.android.network.category
 
-import org.hisp.dhis.android.core.arch.api.fields.internal.Fields
-import org.hisp.dhis.android.core.arch.api.filters.internal.Filter
+import org.hisp.dhis.android.core.arch.api.payload.internal.Payload
 import org.hisp.dhis.android.core.category.Category
 import org.hisp.dhis.android.core.category.internal.CategoryNetworkHandler
 import org.hisp.dhis.android.network.common.HttpServiceClientKotlinx
-import org.hisp.dhis.android.network.common.PayloadJson
 import org.koin.core.annotation.Singleton
 
 @Singleton
@@ -41,12 +39,12 @@ internal class CategoryNetworkHandlerImpl(
 ) : CategoryNetworkHandler {
     private val service: CategoryService = CategoryService(httpClient)
 
-    override suspend fun categories(
-        fields: Fields<Category>,
-        uids: Filter<Category>,
-        paging: Boolean,
-    ): PayloadJson<Category> {
-        val apiPayload = service.getCategories(fields, uids, paging)
+    override suspend fun getCategories(categoryUids: Set<String>): Payload<Category> {
+        val apiPayload = service.getCategories(
+            CategoryFields.allFields,
+            CategoryFields.uid.`in`(categoryUids),
+            paging = false,
+        )
         return apiPayload.mapItems(::categoryDtoToDomainMapper)
     }
 }
